refactor: cleanup event-cache aliases and runtime_support imports (ARCH-B-3 step 5)
Focused verification passed on the final step-5 state:\n- nix-shell --run 'uv run pytest tests/test_event_cache.py tests/test_partial_reply_context.py tests/test_thread_history.py tests/test_thread_mode.py -x -n 0 --no-cov -q'\n- uv run tach check\n\nThe prescribed final pytest command included tests/test_conversation_cache*.py, but that glob matches no files in this repository, so the meaningful equivalent run used the existing tests/test_matrix*.py and tests/test_thread_*.py slices earlier in the session.\n\nRepo-wide pre-commit is still blocked by unrelated existing ty diagnostics outside ARCH-B-3 in:\n- src/mindroom/api/openai_compat.py\n- src/mindroom/response_runner.py\n- tests/test_agno_history.py\n- tests/test_ai_user_id.py
